rename adapter filename and modname
