+++ src/lib.rs
@@ -51,7 +51,7 @@ macro_rules! pixel_game {
         use rust_pixel::util::get_project_path;
 
         #[cfg(target_arch = "wasm32")]
-        use rust_pixel::render::adapter::web::{input_events_from_web, WebAdapter};
+        use rust_pixel::render::adapter::web_adapter::{input_events_from_web, WebAdapter};
         use wasm_bindgen::prelude::*;
         #[cfg(target_arch = "wasm32")]
         use wasm_bindgen_futures::js_sys;
